+++ src/program/Config.cpp
@@ -101,6 +101,7 @@ void Config::save(const std::string& gamepath) {
     settings.setValue("proton_path", proton_path.c_str());
     settings.setValue("editor_autoscroll", editor_autoscroll);
     settings.setValue("editor_rewind_seek", editor_rewind_seek);
+    settings.setValue("editor_rewind_fastforward", editor_rewind_fastforward);
 
     settings.beginGroup("keymapping");
 
@@ -242,6 +243,7 @@ void Config::load(const std::string& gamepath) {
     proton_path = settings.value("proton_path", "").toString().toStdString();
     editor_autoscroll = settings.value("editor_autoscroll", editor_autoscroll).toBool();
     editor_rewind_seek = settings.value("editor_rewind_seek", editor_rewind_seek).toBool();
+    editor_rewind_fastforward = settings.value("editor_rewind_fastforward", editor_rewind_fastforward).toBool();
 
     /* Load key mapping */