+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add the possibility to specify custom monitor resolutions
 * Add a debug native fileio option
 * Skip GL draw function for EGL
+* Add an option to disable fastforward during input editor rewind (#491)
 
 ### Changed
 ### Fixed